Create QUESTION ON ARRAY AND CONTROL STRUCTURES
QUESTION ON ARRAY AND CONTROL STRUCTURES

Q1. Write a C# program that prompt the user to input three numbers .The program should then output the numbers in ascending order.

Q2 .Write a PHP Function ,smallestindex ,that takes as parameters an int array and its size,and returns the index of the smallest element in the array .Also,write a program to test your function .

Q3.Write a C# program that prompts the user to input a string and outputs the string in uppercase(Use a character array to store the string)

Q4. Write a C# program to compute the addition of of N by M matrices.Allow the user to determine the size of the row and column

Q5. Write a C# program that declares an array alpha of 50 components of the type float .Initialize the array so that the first 25 components are equal to the square of the index  variable and the last 25 components are equal to three times the index variable.Output the array so that 10 elements per line are printed.

Q6 Write a C# program that prompts the user to input a number .The program should then output the number and a message saying wether the number is positive ,negative ,or zero.